requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
h11==0.14.0
idna==3.6
ipython==8.23.0
jedi==0.19.1
jsonlines==4.0.0
loguru==0.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nomic==3.0.21
numpy==1.26.4
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uv==0.1.29
uvicorn==0.29.0
wcwidth==0.2.13
